Made some UI/Made some animations
Made Exit mechanism. Made Fade In/Out mechanism and transition for now. Made some Reset mechanism for the restart also. Also made changes to the slide mechanism by changing the speed of the camera parallax without using the velocity of the player. Also fixed Double Jump, changed the counter to 1 as the player is standing on the ground so when the player jumps, it gets reset to 1