Logic Changes/New Animation PlaceHolder
Fire Obstacle was already good, no need to change it. Fixed Effects which are now disabled and renabled accordingly at the time of restart. RockHead only unalives when the player is at the bottom of it. Moving on to making Power-Ups. Made Shield Animation, RockHead/SpikeHead Exploding Animation. Some minor code changes including callbacks for the new power-ups system.